chore: split checkered.rs to surface and spatial checker
